requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SocketIO==5.3.6
Flask-Limiter==3.6.0
Flask-RESTx==1.3.0
python-dotenv==1.0.1
psycopg[binary]==3.2.6
redis==5.0.1
celery==5.3.6
gunicorn==21.2.0
marshmallow==3.21.1
marshmallow-sqlalchemy==1.0.0
PyJWT==2.8.0
bcrypt==4.1.2
python-dateutil==2.8.2
requests==2.31.0
Pillow==11.1.0
boto3==1.34.34
alembic==1.13.1
pytest==8.0.2
pytest-cov==4.1.0
black==24.2.0
flake8==7.0.0
Flask-Admin==1.6.1
